refactor import_samples class
